♻️ refactor: extract board type
