+++ src/app/(with-header-sidebar)/dashboard/[id]/edit/_lib/boardService.ts
@@ -1,5 +1,5 @@
 import axiosInstance from '@/lib/axiosInstance';
-import { UpdateDashboardRequestParams } from '../_types/board';
+import { UpdateDashboardRequestParams } from '@/types/dashboards';
 
 export const getBoard = async (id: string) => {
   try {
