--- src/app/(with-header-sidebar)/dashboard/[id]/edit/_types/board.ts
@@ -1,14 +1,0 @@
-export interface Dashboard {
-  id: number;
-  title: string;
-  color: string;
-  userId: number;
-  createdAt: string;
-  updatedAt?: string;
-  createdByMe: boolean;
-}
-
-export interface UpdateDashboardRequestParams {
-  title: string;
-  color: string;
-}
