+++ src/components/sidebar/Dashboards.tsx
@@ -1,6 +1,6 @@
 import Link from 'next/link';
 import { usePathname } from 'next/navigation';
-import type { Dashboard } from '@/app/(with-header-sidebar)/mydashboard/_types/dashboards';
+import type { Dashboard } from '@/types/dashboards';
 import Image from 'next/image';
 import Button from '../Button';
 import styles from './Dashboards.module.css';
